Move remaining doc-comment parsing logic into 'DocCommentParser' file.
